requirements.txt
discord
pillow
requests
pandas
datetime
pytz
wikipedia
qrcode
pywhatkit
youtube_dl
emoji